requirements.txt
PyYAML==6.0
toml==0.10.2
ujson==5.4.0
